requirements.txt
langgraph>=0.0.30
langchain-openai>=0.0.5
python-dotenv>=1.0.0
pyperclip>=1.8.2
pywin32>=306; platform_system=="Windows"